Se cambia tipo de dati idsalida a Long

+++ src/main/java/com/prueba/entity/Salida.java
@@ -49,7 +49,7 @@ public class Salida implements Serializable {
 	@Column(name = "nidsalida")
 	@GeneratedValue(strategy = GenerationType.SEQUENCE,	generator = "seq_salida")
 	@SequenceGenerator(name = "seq_salida", allocationSize = 10)
-	public Integer idSalida;
+	public Long idSalida;
 	
 	@Column(name = "vcnumdocumento")
 	public String numDocumento;
@@ -103,11 +103,11 @@ public class Salida implements Serializable {
 		this.numDocumento = "SIN-" + String.valueOf(this.idSalida);
 	}
 
-	public Integer getIdSalida() {
+	public Long getIdSalida() {
 		return idSalida;
 	}
 
-	public void setIdSalida(Integer idSalida) {
+	public void setIdSalida(Long idSalida) {
 		this.idSalida = idSalida;
 	}
